--- Scrumban/BusinessLogicLayer/DTO/PriorityDTO.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Scrumban.BusinessLogicLayer.DTO
-{
-    public class PriorityDTO
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-}
