--- Scrumban/ClientApp/src/components/Helper.js
@@ -1,15 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var UserData = /** @class */ (function () {
-    function UserData() {
-        this.userId = 0;
-        this.firstname = "";
-        this.surname = "";
-        this.email = "";
-        this.password = "";
-        this.confirmpassword = "";
-    }
-    return UserData;
-}());
-exports.UserData = UserData;
-//# sourceMappingURL=Helper.js.map
