+++ Scrumban/ClientApp/src/components/Layout.js
@@ -1,31 +1,22 @@
 import React, { Component } from 'react';
 import { Col, Grid, Row } from 'react-bootstrap';
 import { NavMenu } from './NavMenu';
-import { Panel } from './Panel.jsx'
 
 export class Layout extends Component {
   displayName = Layout.name
 
   render() {
     return (
-        <div>
-            <span>
-                <Panel />
-            </span>
-        
-        <Grid fluid>
-              
+      <Grid fluid>
         <Row>
-          <Col sm={1}>
-                        {/*<NavMenu />*/}
+          <Col sm={3}>
+            <NavMenu />
           </Col>
-          <Col sm={10}>
+          <Col sm={9}>
             {this.props.children}
           </Col>
         </Row>
-            </Grid>
-        </div>
-        
+      </Grid>
     );
   }
 }
