+++ Scrumban/ClientApp/src/components/NavMenu.js
@@ -2,22 +2,21 @@
 import { Link } from 'react-router-dom';
 import { Glyphicon, Nav, Navbar, NavItem } from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
-//import './NavMenu.css';
+import './NavMenu.css';
 
 export class NavMenu extends Component {
   displayName = NavMenu.name
 
   render() {
-      return (
-          
-      <Navbar>
+    return (
+      <Navbar inverse fixedTop fluid collapseOnSelect>
         <Navbar.Header>
           <Navbar.Brand>
             <Link to={'/'}>Scrumban</Link>
           </Navbar.Brand>
           <Navbar.Toggle />
         </Navbar.Header>
-        <Navbar>
+        <Navbar.Collapse>
           <Nav>
             <LinkContainer to={'/'} exact>
               <NavItem>
@@ -34,15 +33,9 @@ export class NavMenu extends Component {
                 <Glyphicon glyph='th-list' /> Fetch data
               </NavItem>
             </LinkContainer>
-            <LinkContainer to={'/tasks'}>
-              <NavItem>
-                <Glyphicon glyph='th-list' /> Tasks
-              </NavItem>
-            </LinkContainer>
           </Nav>
-        </Navbar>
+        </Navbar.Collapse>
       </Navbar>
-      
     );
   }
 }
