--- Scrumban/ClientApp/src/components/Panel.css
@@ -1,30 +1,0 @@
-﻿.panelUl {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-    background-color: #333;
-}
-
-.panelLi {
-    float: right;
-    
-}
-.panelBtn {
-    display: block;
-    color: white;
-    text-align: center;
-    padding: 14px 16px;
-    text-decoration: none;
-    background-color: #333;
-    border: none;
-
-
-}
-    .panelBtn:hover, .panelBtn:focus {
-        background-color: #4189C7;
-    }
-    
-
-   
-
