--- Scrumban/ClientApp/src/components/Register.css
@@ -1,16 +1,0 @@
-﻿@media all and (min-width: 480px) {
-    .Register {
-        padding: 60px 0;
-    }
-
-        .Signup form {
-            margin: 10px auto;
-            max-width: 320px;
-        }
-}
-
-.Register form span.help-block {
-    font-size: 14px;
-    padding-bottom: 10px;
-    color: #999;
-}
