--- Scrumban/Models/Entities/Priority.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Scrumban.Models.Entities
-{
-    public class Priority
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-}
