--- Scrumban/Models/Users.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using System.ComponentModel.DataAnnotations;
-
-namespace CustomIdentityApp.Models
-{
-    public class Users
-    {
-        [Key]
-        public int UserID { get; set; }
-        public string FirstName { get; set; }
-        public string Surname { get; set; }
-        public string Email { get; set; }
-        public int RoleID { get; set; }
-        public string Password { get; set; }
-        public int PictureID { get; set; }
-
-    }
-}
