+++ Scrumban/Startup.cs
@@ -1,24 +1,10 @@
- using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SpaServices.ReactDevelopmentServer;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Scrumban.Models;
-using CustomIdentityApp.Models;
-using Microsoft.AspNet.OData.Extensions;
-using Microsoft.AspNet.OData.Builder;
-using Microsoft.OData.Edm;
-using Scrumban.Models.Entities;
-using Scrumban.DataAccessLayer;
-using Scrumban.DataAccessLayer.Interfaces;
-using Scrumban.DataAccessLayer.Repositories;
-using Scrumban.BusinessLogicLayer;
-using Scrumban.BusinessLogicLayer.Interfaces;
-//using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace Scrumban
 {
@@ -31,32 +17,19 @@ namespace Scrumban
 
         public IConfiguration Configuration { get; }
 
-      
+        // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            string connection = Configuration.GetConnectionString("DefaultConnection");
-            
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+
+            // In production, the React files will be served from this directory
             services.AddSpaStaticFiles(configuration =>
             {
                 configuration.RootPath = "ClientApp/build";
             });
-            
-            services.AddODataQueryFilter();
-            services.AddOptions();
-            services.AddTransient<ITaskService, TaskService>();
-            services.AddTransient<IUnitOfWork, UnitOfWork>();
-            services.AddTransient<IRepository<Task>, TaskRepository>();
-           
-            services.AddDbContext<ScrumbanContext>(options =>
-                options.UseSqlServer(connection));
-            
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-            services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-            services.AddOData();
         }
 
-       
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
             if (env.IsDevelopment())
@@ -73,17 +46,11 @@ namespace Scrumban
             app.UseStaticFiles();
             app.UseSpaStaticFiles();
 
-            app.UseMvc(routeBuilder =>
-            {
-                routeBuilder.EnableDependencyInjection();
-                routeBuilder.Filter().OrderBy().Count().Expand().Select();
-            });
-
             app.UseMvc(routes =>
             {
                 routes.MapRoute(
                     name: "default",
-                    template: "{controller}/{action=index}/{id?}");
+                    template: "{controller}/{action=Index}/{id?}");
             });
 
             app.UseSpa(spa =>
